Fixed Classes

+++ PEExecutionEngine.h
@@ -6,7 +6,7 @@ uint64_t roundUp(uint64_t numToRound, uint64_t multiple)
     if (multiple == 0)
         return numToRound;
 
-    int remainder = numToRound % multiple;
+    uint64_t remainder = numToRound % multiple;
     if (remainder == 0)
         return numToRound;
